requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
aiofiles==23.2.1

# HTTP requests and system monitoring
requests==2.31.0
psutil==5.9.6

# Image processing for background generation
Pillow==10.1.0

# QR code generation
qrcode[pil]==7.4.2

# Web automation for webcast feature
playwright==1.40.0

# Optional: Enhanced logging and configuration
python-json-logger==2.0.7

# Development and testing dependencies (uncomment if needed)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2  # For testing FastAPI endpoints

# Production deployment (uncomment if needed)
# gunicorn==21.2.0
# supervisor==4.2.5